+++ src/app/app.component.html
@@ -1,4 +1,4 @@
-<app-navbar (changeView)="onChangeView($event)"></app-navbar>
+<app-navbar></app-navbar>
 <div class="main container-fluid flex">
     <div class="row flex">
         <div class="col-sm-12 col-md-7 order-sm-first order-md-last map-container" [ngClass]="{'hide-map' : pressed}">
