+++ src/app/category-bar-calendar/category-bar-calendar.component.ts
@@ -91,5 +91,5 @@ export class CategoryBarCalendarComponent implements OnInit {
   private objectKeys(obj) {
     return Object.keys(obj);
   }
-  
+
 }
