+++ src/app/category.service.ts
@@ -9,14 +9,14 @@ const httpOptions = { headers: new HttpHeaders({ 'Content-Type': 'application/js
 @Injectable()
 export class CategoryService {
 
+  // CONSTRUCTOR
+
+  constructor(private http: HttpClient) { }
+
   // private apiUrl = "https://www.mappening.io/api/v1/events/event-categories";
   // private apiUrl = "http://0.0.0.0:5000/api/v2/events/categories";
   private apiUrl = "https://www.mappening.io/api/v2/events/categories";
   private selectedCategory = "all";
-
-  // CONSTRUCTOR
-
-  constructor(private http: HttpClient) { }
 
   getSelectedCategory(){
     return this.selectedCategory;
